refactor  : queryclient server-side fetch
